feat: 메인 페이지 컴포넌트 리팩토링 및 스켈레톤 추가, SSR 전환
