+++ src/components/main/Top5.tsx
@@ -1,115 +1,40 @@
 "use client"
 
 import { cn } from "@/lib/utils"
-import { useApi } from "@/hooks/useApi"
 import { MainProject } from "@/lib/projectInterface"
 import { useEffect, useState } from "react"
-import { useRouter } from "next/navigation";
+import { useRouter } from "next/navigation"
 import ProjectOverlayItem from "../common/ProjectOverlayItem"
 import { SecondaryButton } from "../common/Button"
+import { CategoryItem } from "@/app/page"
+import Top5Skeleton from "./Top5Skeleton";
+import dynamic from "next/dynamic"
 
-interface Response {
-  allTopView: MainProject[],
-  categoryTopView: CategoryItem[]
-}
-
-interface CategoryItem {
-  categoryId: number,
-  categoryName: string,
-  items: MainProject[]
-}
-
-export default function TopFive() {
+function TopFiveComponent({ allProjects, categoryProjects }: { allProjects: MainProject[], categoryProjects: CategoryItem[] }) {
   const router = useRouter()
-  const { isLoading, apiCall } = useApi()
-  const [activeTab, setActiveTab] = useState(0) 
-  const [categories, setCategories] = useState<{ [key: number]: string }>({
-    0: "전체"
-  })
-  const [categoryProjects, setCategoryProjects] = useState<{ [key: number]: MainProject[] }>({
-    0: []
-  })
-  const [hasError, setHasError] = useState(false)
-  const rank1Project = categoryProjects[activeTab] ? categoryProjects[activeTab][0] : null;
-  const restProject = categoryProjects[activeTab] ? categoryProjects[activeTab].slice(1) : [];
-  const top5Projects = categoryProjects[activeTab] || [];
+  const [activeTab, setActiveTab] = useState(0)
+  const [projects, setProjects] = useState<MainProject[]>(allProjects);
+  const [isClient, setIsClient] = useState(false);
 
-  const loadProjects = () => {
-    setHasError(false);
-    apiCall<Response>("/api/project/categoryRankingView", "GET").then(({ data, error }) => {
-      if (error) {
-        console.error('Failed to load category ranking view:', error);
-        setHasError(true);
-        return;
-      }
-      if (data?.allTopView && data.categoryTopView) {
-        setCategoryProjects({
-          ...categoryProjects,
-          0: data?.allTopView
-        })
-        data.categoryTopView.map(categoryItem => {
-          setCategories(prev => {
-            prev[categoryItem.categoryId] = categoryItem.categoryName
-            return prev
-          })
-          setCategoryProjects(prev => {
-            prev[categoryItem.categoryId] = categoryItem.items
-            return prev
-          })
-        })
-      }
-      
-    })
-  }
+  const rank1Project = projects[0] ? projects[0] : null;
+  const restProject = projects.slice(1) || [];
 
   const moveToPage = (categoryId: number) => {
     if (categoryId) router.push(`/category/${categoryId}`)
   }
 
-  const moveToProjectPage = (projectId: number) => {
-    router.push(`/project/${projectId}`)
-  }
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
 
   useEffect(() => {
-    loadProjects();
-  }, [])
+    setProjects(activeTab === 0 ? allProjects : categoryProjects[activeTab].items)
+  }, [activeTab, allProjects, categoryProjects])
 
-  // 에러가 있으면 아무것도 렌더링하지 않음
-  if (hasError) {
-    return null;
-  }
-
-  if (isLoading || top5Projects.length === 0) {
+  // 클라이언트 사이드에서만 렌더링
+  if (!isClient) {
     return (
-      <div className="w-full max-w-7xl mx-auto px-4 py-12" data-cy="category-top5-skeleton">
-        <div className="flex items-center mb-8">
-          <h2 className="text-2xl font-bold">
-            BEST <span className="text-main-color">TOP 5</span> <span className="text-2xl">🏆</span>
-          </h2>
-        </div>
-        <div className="flex justify-between mb-8">
-          <div className="flex flex-wrap gap-2">
-            {[0,1,2].map(i => (
-              <div key={i} className="px-4 py-2 rounded-full bg-gray-200 animate-pulse w-20 h-8" />
-            ))}
-          </div>
-          <div className="px-4 py-2 rounded bg-gray-200 animate-pulse w-32 h-8" />
-        </div>
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-          {/* 1등 (큰 카드) */}
-          <div className="flex flex-col items-start animate-pulse w-full">
-            <div className="w-full aspect-square bg-gray-200 rounded-lg" />
-          </div>
-          {/* 2~5등 (2x2 그리드) */}
-          <div className="grid grid-cols-2 grid-rows-2 gap-4">
-            {[0,1,2,3].map(i => (
-              <div key={i} className="flex flex-col items-start animate-pulse w-full">
-                <div className="w-full aspect-square bg-gray-200 rounded-lg" />
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
+      <Top5Skeleton />
     );
   }
   
@@ -124,7 +49,7 @@ export default function TopFive() {
       {/* Category filters */}
       <div className="flex justify-between mb-8">
       <div className="flex flex-wrap gap-2">
-        {Object.keys(categories).map((categoryId) => (
+        {Object.keys(categoryProjects).map((categoryId) => (
           <button
             key={categoryId}
             className={cn(
@@ -133,7 +58,7 @@ export default function TopFive() {
             )}
             onClick={() => setActiveTab(+categoryId)}
           >
-            {categories[+categoryId]}
+            {categoryProjects[+categoryId].categoryName}
           </button>
         ))}
       </div>
@@ -163,7 +88,7 @@ export default function TopFive() {
 
         {/* Regular projects (smaller) - takes up the other half in a 2x2 grid */}
         {restProject && <div className="grid grid-cols-2 grid-rows-2 gap-4">
-          {restProject.map((project, idx) => 
+          {restProject.map((project: MainProject) => 
             <ProjectOverlayItem
             key={project.projectId}
             className="w-full"
@@ -185,3 +110,17 @@ export default function TopFive() {
     </div>
   )
 }
+
+const TopFive = dynamic(
+  () => Promise.resolve(({ allProjects, categoryProjects }: { allProjects: MainProject[], categoryProjects: CategoryItem[] }) => {
+    return (
+      <TopFiveComponent allProjects={allProjects} categoryProjects={categoryProjects} />
+    )
+  }),
+  {
+    ssr: false,
+    loading: () => <Top5Skeleton />
+  }
+)
+
+export default TopFive
